+++ CHANGELOG.md
@@ -38,6 +38,9 @@ x.x.x Release notes (yyyy-MM-dd)
   removed.
 * The class methods `Object.className()`, `Object.objectUtilClass()`, and
   the property `Object.isInvalidated` can no longer be overriden.
+* Fix empty strings and null on the right side of `BEGINSWITH`, `ENDSWITH` and
+  `CONTAINS` operators in predicates to match Foundation's semantics of never
+  matching any strings.
 
 ### Enhancements